requirements.txt
django~=3.1.6
psycopg2-binary~=2.8.6
python-dotenv~=0.17.0
djangorestframework~=3.12.0
whitenoise~=5.2.0